Tiny changes

+++ cpp_module_07/ex01/iter.hpp
@@ -16,8 +16,8 @@
 # include <iostream>
 
 template< typename T >
-void	iter(T *arr, size_t size, void (*funct_ptr)(T const &val)) {
-	for (size_t i = 0; i < size; i++) {
+void	iter(T *arr, int size, void (*funct_ptr)(T const &val)) {
+	for (int i = 0; i < size; i++) {
 		(*funct_ptr)(arr[i]);
 	}
 }
